De-singletonize the store
Create the store once in the main `index.js` and then inject it where
needed.  This accomplishes three main goals:

1. Gets rid of a singleton
2. Makes our code look more like the official Redux examples
3. Supports server-side rendering should any project based on the
boilerplate need that.  With server rendering, you create a new store
on the server for every request, so you need to be able to inject that
into the component tree.  A singleton just wouldn’t cut it there.
