+++ client/routes/AppRouter.js
@@ -1,17 +1,19 @@
-import React from 'react'
-import { browserHistory, Router } from 'react-router'
-import { syncHistoryWithStore } from 'react-router-redux'
+import React, { PropTypes } from 'react'
+import { Router } from 'react-router'
 import { FixNamedRoutesSupport as allowNames } from 'react-router-named-routes'
 
 import routes from 'routes'
-import store from 'store'
 
 allowNames(routes)
 
-const history = syncHistoryWithStore(browserHistory, store)
+const propTypes = {
+  history: PropTypes.object.isRequired
+}
 
-export default function AppRouter() {
+export default function AppRouter({ history }) {
   return (
     <Router children={routes} history={history} />
   )
 }
+
+AppRouter.propTypes = propTypes
